+++ attribution/src/main/java/com/affise/attribution/network/entity/PostBackModel.kt
@@ -11,3 +11,17 @@ data class PostBackModel(
     val metrics: List<SerializedEvent>? = null,
     val internalEvents: List<SerializedEvent>? = null,
 )
+
+internal fun PostBackModel.asFirstOpen(): PostBackModel = PostBackModel(
+    parameters = this.parameters.entries.associate {
+        return@associate if (it.key == ProviderType.INSTALL_FIRST_EVENT) {
+            it.key to true
+        } else {
+            it.key to it.value
+        }
+    },
+    events = this.events,
+    logs = this.logs,
+    metrics = this.metrics,
+    internalEvents = this.internalEvents,
+)