+++ attribution/src/main/java/com/affise/attribution/usecase/IndexUseCase.kt
@@ -1,0 +1,6 @@
+package com.affise.attribution.usecase
+
+interface IndexUseCase {
+    fun getUuidIndex(): Long
+    fun getAffiseEventIdIndex(): Long
+}